+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [unreleased]
+
+### Added
+
+- Support for positionals with variable amount of values (#1).
+
 ## [0.1.1] - 2025-06-18
 
 ### Changed
@@ -10,5 +16,6 @@
 
 Initial release.
 
+[unreleased]: https://github.com/phoenixr-codes/powercli/compare/v0.1.1...HEAD
 [0.1.1]: https://github.com/phoenixr-codes/powercli/compare/v0.1.0...v0.1.1
 [0.1.0]: https://github.com/phoenixr-codes/powercli/releases/tag/v0.1.0